oldverybad_implementation_do_not_use_and_do_not_look/requirements.txt
aiohttp>=3.8.0,<4.0.0
aiosql>=8.0,<10.0
asyncpg>=0.28.0,<0.30.0
redis>=4.5.0,<7.0.0
bcrypt>=4.0.0,<5.0.0
PyJWT>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
Jinja2>=3.1.0,<4.0.0
aiofiles>=23.0.0,<24.0.0
croniter>=2.0.0,<3.0.0